Season-3/GAN/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
datasets>=2.14.0
transformers>=4.30.0
Pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
tensorboard>=2.13.0
huggingface-hub>=0.16.0
accelerate>=0.20.0
